Seperate some al files out like they seem to be in ultralib
