--- lib/src/al/global_asm/alSynStartVoice.c
@@ -1,7 +1,0 @@
-/* The comment below is needed for this file to be picked up by generate_ld */
-/* RAM_POS: 0x800C96F0 */
-
-#include "types.h"
-#include "macros.h"
-
-GLOBAL_ASM("lib/asm/non_matchings/unknown_0CA2F0/alSynStartVoice.s")
